4-mqtt_2hosts/pyproject.toml
[project]
name = "4-mqtt-2hosts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
]
